add instead of verify
